qemu: remove 0.9.x
QEMU 0.9.x is obsolete this days and with all modifications made for
QEMU 0.10.x in OE, versions 0.9.x are unusable anyway. 0.10.3 was
introduced 10 month ago, so a good transition time was also given.
Now it's time to just kill it.

Also remove cvs and svn versions since those are 0.9.x leftovers and
QEMU moved to git long ago.

Also remove gcc3 checks since that are not relevant for QEMU 0.10+.

Also remove from icecc blacklist, since QEMU builds fine with it and
the reason for blacklisting was GCC 3.

Signed-off-by: Roman I Khimov <khimov@altell.ru>
Acked-by: Tom Rini <tom_rini@mentor.com>
Acked-by: Koen Kooi <koen@openembedded.org>
Acked-by: Marcin Juszkiewicz <marcin@juszkiewicz.com.pl>